GPII-1993 : First attempt porting our current provisioning/build.bat to Powershell.
Created a new Powershell module "Provisioning.psm1" for containing all
the needed function and aliases for building GPII in any Windows
version. Create new Powershell script "Build.ps1" mimicking the
behaviour of build.bat.

Currently we are requiring Powershell => 3.0, in the future if we want
to support lower Windows version we will need to downgrade to Powershell
2.0.