GPII-442:  Normalized value of argv[0]
The result returned by system calls varies for the command line first
parameter, or argv[0].  Sometimes the result is simply the name of the
command, for example, "Notepad". Or it might be the full path, and may
or may not include the ".exe" extension.  Modified the code to behave
the way the node process presents itself, namely, to use the full
path (when available).